Implement updateStoryListDataStores function in AppDataContext for improved story list management; refactor ServerStoryList to utilize this new function, enhancing data persistence and refresh handling.
