+++ src/features/storyList/ServerStoryList.tsx
@@ -1,18 +1,11 @@
 import { createAsync } from "@solidjs/router";
-import {
-  createEffect,
-  createMemo,
-  createRenderEffect,
-  untrack,
-} from "solid-js";
+import { createEffect, createMemo, untrack } from "solid-js";
 import { createStore, reconcile } from "solid-js/store";
-import { isServer } from "solid-js/web";
 
 import {
+  updateStoryListDataStores,
   useAppData,
   useDataStore,
-  useMessagesStore,
-  useRefreshStore,
 } from "~/contexts/AppDataContext";
 import { mapStoriesToSummaries } from "~/lib/getSummaryViaFetch";
 import { TopStoriesType } from "~/models/interfaces";
@@ -21,23 +14,14 @@ import { getStoryListByType } from "~/server/queries";
 import { HnStoryListBody } from "./HnStoryListBody";
 import { HnStoryListToggle } from "./HnStoryListToggle";
 
-import type { HnStorySummary, StoryPage } from "~/models/interfaces";
+import type { HnStorySummary } from "~/models/interfaces";
 
 export function ServerStoryList(props: { page: TopStoriesType }) {
-  console.log("*** ServerStoryPage", { isServer, page: props.page });
   const isClientMounted = useAppData().isClientMounted;
   const dataStore = useDataStore();
-  const refreshStore = useRefreshStore();
-  const messagesStore = useMessagesStore();
 
   const data = createAsync(async () => {
     const isClient = untrack(() => isClientMounted());
-
-    console.log("*** getStoryListByType query", {
-      isServer,
-      isClient,
-      page: props.page,
-    });
 
     if (isClient) {
       const list = await dataStore.getContentForPage(props.page);
@@ -62,41 +46,15 @@ export function ServerStoryList(props: { page: TopStoriesType }) {
   );
 
   createEffect(() => {
-    // in theory this store should give a nice reorder effect for auto-animate?
     const sums = summaries();
-    if (!sums) {
-      return;
+    if (sums) {
+      // in theory this store should give a nice reorder effect for auto-animate?
+      setSummaryStore(reconcile(sums, { key: "id" }));
     }
-    setSummaryStore(reconcile(sums));
   });
-
-  // TODO: persistence should move out of the render path
 
   createEffect(() => {
-    const p = props.page as StoryPage;
-    const latest = data.latest;
-    const d = latest?.result;
-    if (!latest?.startedFromServer || !d || !Array.isArray(d)) {
-      return;
-    }
-
-    console.log("*** persisting story list", p, d.length);
-    void dataStore.persistStoryList(p, d);
-  });
-
-  createRenderEffect(() => {
-    dataStore.setRefreshType({
-      type: "storyList",
-      page: props.page as StoryPage,
-    });
-    messagesStore.addMessage("refresh", "setRefreshType", { page: props.page });
-  });
-
-  createRenderEffect(() => {
-    const page = props.page as StoryPage;
-    if (refreshStore.refreshRequestedTimestamps[page] === undefined) {
-      refreshStore.setRefreshRequestedTimestamp(page);
-    }
+    updateStoryListDataStores(props.page, data);
   });
 
   return (
